OneDrive/Documents/document_reasoning_assistant/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
pymupdf>=1.23.0
python-docx>=0.8.11
pymongo>=4.6.0
requests>=2.31.0
pinecone-client>=2.2.4
groq>=0.4.1
